back/requirements.txt
Flask
Werkzeug
python-dotenv
Flask-SQLAlchemy
PyMySQL
marshmallow
PyJWT
Flask-Cors
